MainContainer/MoviesList names/body correction
